+++ idea/frontend/src/entities/message/model/types.ts
@@ -1,4 +1,4 @@
-import { IProgram } from 'entities/program';
+import { IProgram } from 'features/program';
 import { IBase } from 'shared/types';
 
 import { Type, EntryPoint, ReadReason } from './consts';
