--- idea/frontend/src/entities/program/index.ts
@@ -1,6 +1,0 @@
-import { IProgram, ProgramStatus, PROGRAM_STATUS_NAME } from './model';
-import { getBulbStatus } from './helpers';
-import { HorizontalProgramCard } from './ui/horizontalProgramCard';
-
-export { ProgramStatus, PROGRAM_STATUS_NAME, HorizontalProgramCard, getBulbStatus };
-export type { IProgram };
