--- idea/frontend/src/entities/program/model/index.ts
@@ -1,5 +1,0 @@
-import { IProgram } from './types';
-import { PROGRAM_STATUS_NAME, ProgramStatus } from './consts';
-
-export { ProgramStatus, PROGRAM_STATUS_NAME };
-export type { IProgram };
