--- idea/frontend/src/entities/program/ui/horizontalProgramCard/index.ts
@@ -1,3 +1,0 @@
-import { HorizontalProgramCard } from './HorizontalProgramCard';
-
-export { HorizontalProgramCard };
