+++ idea/frontend/src/features/program/consts.ts
