+++ idea/frontend/src/features/program/index.ts
@@ -1,3 +1,6 @@
-import { ProgramTable } from './ui';
+import { ProgramTable, ProgramCard, ProgramsSearch } from './ui';
+import { IProgram } from './types';
+import { ProgramStatus } from './consts';
 
-export { ProgramTable };
+export { ProgramTable, ProgramCard, ProgramStatus, ProgramsSearch };
+export type { IProgram };
