+++ idea/frontend/src/features/program/ui/index.ts
@@ -1,3 +1,5 @@
 import { ProgramTable } from './program-table';
+import { ProgramCard } from './program-card';
+import { ProgramsSearch } from './programs-search';
 
-export { ProgramTable };
+export { ProgramTable, ProgramCard, ProgramsSearch };
