+++ idea/frontend/src/features/program/ui/program-card/index.ts
@@ -1,0 +1,3 @@
+import { ProgramCard } from './program-card';
+
+export { ProgramCard };
