+++ idea/frontend/src/features/program/ui/programs-search/index.ts
@@ -1,0 +1,3 @@
+import { ProgramsSearch } from './programs-search';
+
+export { ProgramsSearch };
