+++ idea/frontend/src/features/program/ui/programs-search/programs-search.module.scss
