+++ idea/frontend/src/hooks/usePrograms.ts
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useAlert } from '@gear-js/react-hooks';
 
 import { FetchProgramsParams, ProgramPaginationModel } from 'api/program/types';
-import { IProgram } from 'entities/program';
+import { IProgram } from 'features/program';
 import { DEFAULT_LIMIT } from 'shared/config';
 import { fetchPrograms, getLocalPrograms } from 'api';
