+++ idea/frontend/src/pages/programs/ui/programsPage/ProgramsPage.tsx
@@ -4,12 +4,11 @@ import { useAccount } from '@gear-js/react-hooks';
 
 import { OwnerFilter } from 'api/consts';
 import { usePrograms, useDataLoading } from 'hooks';
-import { ProgramStatus } from 'entities/program';
+import { ProgramStatus, ProgramsSearch } from 'features/program';
 
 import { RequestParams } from '../../model/types';
 import { DEFAULT_REQUEST_PARAMS, DEFAULT_FILTER_VALUES } from '../../model/consts';
 import { ProgramsList } from '../programsList';
-import { SearchSettings } from '../searchSettings';
 
 import styles from './ProgramsPage.module.scss';
 
@@ -67,7 +66,7 @@ const ProgramsPage = () => {
           main
         />
       </section>
-      <SearchSettings
+      <ProgramsSearch
         initQuery={params.query || ''}
         isLoggedIn={isLoggedIn}
         initialValues={initialValues}
