--- idea/frontend/src/pages/programs/ui/searchSettings/index.ts
@@ -1,3 +1,0 @@
-import { SearchSettings } from './SearchSettings';
-
-export { SearchSettings };
